+++ src/containerapp/HISTORY.rst
@@ -4,10 +4,11 @@ Release History
 ===============
 upcoming
 ++++++
-* 'az containerapp function show/list': New commands to list and show functions in container apps.
+* 'az containerapp function list': List functions in a container app
+* 'az containerapp function show': Show specific function in a container app
 * 'az containerapp function keys show': Show specific function key in a container app
 * 'az containerapp function keys list': List function keys in a container app
-* 'az containerapp function keys set': Set/Update specific function key in a container app
+* 'az containerapp function keys set': Create a new or update an existing function key in a container app
 * 'az containerapp function invocations summary': Get function invocation summary from Application Insights
 * 'az containerapp function invocations traces': Get function invocation traces from Application Insights
 * 'az containerapp update/up': Disallow changing `--revisions-mode` to Labels.